security(render): harden renderInline html and link tokens
renderInline's html token is now escaped via escapeHtml so model-authored
raw HTML renders as literal text, and link-token hrefs pass through
safeInlineLinkHref (added to src/shared/webview-nav-guard.ts), which shares
ALLOWED_URL_SCHEMES with safeExternalUrlString so the renderInline and openUrl
paths never drift on which absolute schemes are allowed. Allowed: http/https/
mailto plus scheme-less refs (relative paths, anchors); blocked: javascript,
data, vbscript, file, ftp, vscode. Per upstream #65 maintainer steer: escape
(not DOMPurify) and reuse the existing openUrl allowlist rather than adding a
second one.

Bundle boundary: webview-nav-guard.ts is imported by both esbuild bundles
(extension-host platform node, webview platform browser), so it stays
browser-safe (no node: imports). safeWorkspaceFilePath (the openFile guard,
which needs node:path) is split into a new src/shared/workspace-path-guard.ts
imported only by the extension-host; without the split, routing renderInline to
webview-nav-guard.ts would pull node:path into the browser webview bundle and
break that build.
